feat(core): move matches/closest onto Element via a Queryable contract
Extract the CSS-like query engine from group.ts into core/query.ts, typed
against a new structural Queryable interface that both Element and Group
implement (containers additionally expose children/graph for combinator
traversal). matches() and closest() are now methods on every Element -
they apply to any element, not just groups - and Group inherits them
instead of overriding with casts. Free-function exports (query, queryAll,
matches, closest) are unchanged and now come from core/query.

Docs updated to state the methods exist on every element, and query tests
extended to cover leaf-element matches/closest across contexts.
