+++ packages/core/src/core/group.ts
@@ -6,15 +6,20 @@ import {
 } from './element';
 
 import {
+    query,
+    queryAll,
+} from './query';
+
+import type {
+    Queryable,
+} from './query';
+
+import {
     getContainingBox,
 } from '../math';
 
 import {
     OneOrMore,
-    typeIsBoolean,
-    typeIsNil,
-    typeIsNumber,
-    typeIsString,
     valueOneOrMore,
 } from '@ripl/utilities';
 
@@ -25,175 +30,6 @@ import type {
 /** Options for constructing a group, extending element options with an optional initial set of children. */
 export interface GroupOptions extends ElementOptions {
     children?: OneOrMore<Element>;
-}
-
-const ELEMENT_PATTERNS = {
-    id: /#/,
-    class: /\./,
-    name: /([a-zA-Z\d\-_]+)/,
-};
-
-const COMBINATOR_PATTERNS = {
-    child: />/,
-    nextSibling: /\+/,
-};
-
-const QUERY_PATTERNS = {
-    type: new RegExp(`^${ELEMENT_PATTERNS.name.source}`),
-    id: new RegExp(`${ELEMENT_PATTERNS.id.source}${ELEMENT_PATTERNS.name.source}`),
-    class: new RegExp(`${ELEMENT_PATTERNS.class.source}${ELEMENT_PATTERNS.name.source}`, 'g'),
-    attribute: new RegExp(`(\\[${ELEMENT_PATTERNS.name.source}="([^"]*)"\\])+`, 'g'),
-    combinators: new RegExp(`(\\s(?:${Object.values(COMBINATOR_PATTERNS).map(pattern => pattern.source).join('|')})\\s|\\s)`),
-};
-
-const COMBINATOR_PRODUCERS = [
-    {
-        pattern: /^\s$/,
-        produce: element => isGroup(element) ? element.graph(true) : [],
-    },
-    {
-        pattern: COMBINATOR_PATTERNS.child,
-        produce: element => isGroup(element) ? element.children : [],
-    },
-    {
-        pattern: COMBINATOR_PATTERNS.nextSibling,
-        produce: element => element.parent?.children.slice(element.parent.children.indexOf(element) + 1) ?? [],
-    },
-] as {
-    pattern: RegExp;
-    produce: (element: Element) => Element[];
-}[];
-
-function serialiseAttribute(value: unknown) {
-    if (typeIsString(value)) {
-        return value;
-    }
-
-    if (typeIsNumber(value) || typeIsBoolean(value)) {
-        return String(value);
-    }
-
-    try {
-        return JSON.stringify(value).replaceAll(/(^"|"$)/g, '');
-    } catch {
-        return String(value);
-    }
-}
-
-/** Tests whether a single (combinator-free) selector segment matches the element. */
-function matchesSegment(element: Element, segment: string): boolean {
-    const type = segment.match(QUERY_PATTERNS.type)?.at(0);
-    const id = segment.match(QUERY_PATTERNS.id)?.at(0);
-    const classes = Array.from(segment.matchAll(QUERY_PATTERNS.class), match => match.at(0));
-    const attributes = Array.from(segment.matchAll(QUERY_PATTERNS.attribute), match => match.at(0));
-
-    const typeMatch = !type || element.type === type;
-    const idMatch = !id || element.id === id.replace(ELEMENT_PATTERNS.id, '');
-
-    const classMatch = !classes.length || classes.every(cls => {
-        return !!cls && element.classList.has(cls.replace(ELEMENT_PATTERNS.class, ''));
-    });
-
-    const attrsMatch = !attributes.length || attributes.every(attr => {
-        if (!attr) {
-            return false;
-        }
-
-        const [
-            ,
-            ,
-            key,
-            value,
-        ] = Array.from(attr.matchAll(QUERY_PATTERNS.attribute)).at(0) || [];
-
-        return !typeIsNil(key)
-            && !typeIsNil(value)
-            && key in element
-            && serialiseAttribute((element as unknown as Record<string, unknown>)[key]) === value;
-    });
-
-    return typeMatch
-        && idMatch
-        && classMatch
-        && attrsMatch;
-}
-
-function executeQuery(elements: Element[], segments: string[], segmentIndex: number = 0) {
-    if (segmentIndex >= segments.length || !elements.length) {
-        return elements;
-    }
-
-    const segment = segments[segmentIndex];
-
-    if (!segment) {
-        return executeQuery(elements, segments, segmentIndex + 1);
-    }
-
-    if (QUERY_PATTERNS.combinators.test(segment)) {
-        const producer = COMBINATOR_PRODUCERS.find(({ pattern }) => pattern.test(segment));
-
-        if (!producer) {
-            throw new Error('Failed to query!');
-        }
-
-        return executeQuery(elements.flatMap(element => producer.produce(element)), segments, segmentIndex + 1);
-    }
-
-    return executeQuery(elements.filter(element => matchesSegment(element, segment)), segments, segmentIndex + 1);
-}
-
-/** Queries all elements matching a CSS-like selector across the given element(s) and their descendants. */
-export function queryAll<TElement extends Element = Element>(elements: OneOrMore<Element | Group>, selector: string) {
-    const els = ([] as Element[]).concat(elements).flatMap(element => {
-        return isGroup(element) ? element.graph(true) : [element];
-    });
-
-    const normalisedSelector = selector.trim().replace(/\s+/g, ' ');
-
-    return executeQuery(els, normalisedSelector.split(QUERY_PATTERNS.combinators)) as TElement[];
-}
-
-/** Returns the first element matching a CSS-like selector, or `undefined` if none match. */
-export function query<TElement extends Element = Element>(elements: OneOrMore<Element | Group>, selector: string) {
-    return queryAll<TElement>(elements, selector).at(0);
-}
-
-/** Tests whether an element matches a CSS-like selector. */
-export function matches(element: Element, selector: string): boolean {
-    const segments = selector.trim().replace(/\s+/g, ' ').split(QUERY_PATTERNS.combinators).filter(Boolean);
-
-    if (!segments.length) {
-        return false;
-    }
-
-    // A single simple segment can be tested directly; multi-segment selectors (with combinators)
-    // are resolved against the element's root so ancestor/sibling constraints are honoured.
-    if (segments.length === 1) {
-        return matchesSegment(element, segments[0]);
-    }
-
-    let root: Element = element;
-
-    while (root.parent) {
-        root = root.parent;
-    }
-
-    return queryAll(root, selector).includes(element);
-}
-
-/** Returns the closest ancestor (including the element itself) matching a CSS-like selector, or `undefined`. */
-export function closest<TElement extends Element = Element>(element: Element, selector: string): TElement | undefined {
-    let current: Element | undefined = element;
-
-    while (current) {
-        if (matches(current, selector)) {
-            return current as TElement;
-        }
-
-        current = current.parent;
-    }
-
-    return undefined;
 }
 
 /** Type guard that checks whether a value is a `Group` instance. */
@@ -207,7 +43,7 @@ export function createGroup(...options: ConstructorParameters<typeof Group>) {
 }
 
 /** A container element that manages child elements, providing scenegraph traversal, CSS-like querying, and composite bounding boxes. */
-export class Group<TEventMap extends ElementEventMap = ElementEventMap> extends Element<BaseElementState, TEventMap> {
+export class Group<TEventMap extends ElementEventMap = ElementEventMap> extends Element<BaseElementState, TEventMap> implements Queryable {
 
     #elements = new Set<Element>();
 
@@ -300,16 +136,6 @@ export class Group<TEventMap extends ElementEventMap = ElementEventMap> extends 
     /** Returns all descendants matching the CSS-like selector. */
     public queryAll<TElement extends Element = Element>(selector: string) {
         return queryAll<TElement>(this, selector);
-    }
-
-    /** Tests whether this group matches the CSS-like selector. */
-    public matches(selector: string): boolean {
-        return matches(this as unknown as Element, selector);
-    }
-
-    /** Returns the closest ancestor (including this group) matching the CSS-like selector, or `undefined`. */
-    public closest<TElement extends Element = Element>(selector: string) {
-        return closest<TElement>(this as unknown as Element, selector);
     }
 
     /** Finds a descendant element by its unique id, or `undefined` if none match. */
